requirements.txt
streamlit
pandas
pillow
numpy
scikit-learn